+++ bfabric_app_runner/src/bfabric_app_runner/cli/cmd_prepare.py
@@ -73,7 +73,7 @@ def _write_workunit_makefile(path: Path, app_ref: Path) -> None:
 
 def _write_app_env_file(
     path: Path,
-    app_ref: Path | str,
+    app_ref: Path,
     workunit_ref: Path,
     ssh_user: str | None,
     force_storage: Path | None,