+++ bfabric_app_runner/docs/changelog.md
@@ -9,9 +9,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - `bfabric-cli prepare workunit` does not accept module refs anymore and will resolve the app spec path.
 - `bfabric-cli prepare workunit --force-app-version` to force a specific app version for the workunit.
 
+### Removed
+
+- Removed experimental `bfabric-app-runner deploy build-app-zip` command and associated app.zip functionality. This functionality has been superseded by the more robust `CommandPythonEnv` approach for Python environment management.
+- Removed module path support for app specification. `CommandPythonEnv` is cleaner than the workflow this tried to enable.
+
 ### Added
 
 - `copier` based template/demo application for development and end-to-end testing of bfabric-app-runner.
+- Added `ResolvedDirectory` type to represent directories resolved from resource archives.
+- Added `BfabricResourceArchiveSpec` to specify input archives which should be extracted  (and select which files are needed).
+- Validation logic has been added for `ResolvedDirectory` and `BfabricResourceArchiveSpec`. In particular a `ResolvedDirectory` may never overlap with a `ResolvedFile` or `ResolvedStaticFile` path.
 
 ## \[0.1.2\] - 2025-07-08